tests/wpt/tests/css/css-sizing/aspect-ratio/replaced-element-028.html

<!DOCTYPE html>
<title>CSS aspect-ratio: img</title>
<link rel="author" title="Google LLC" href="https://www.google.com/">
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#aspect-ratio">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<img id="test" src="error.png" style="width: 100px; aspect-ratio: 1/5; background: yellow;" alt="Should not be 500px high">

<script>
function run_test() {
  test(function() {
    var img = document.getElementById("test");
    assert_not_equals(img.offsetHeight, 500);
  });
}
onload = run_test;
</script>
